refactor(analysis): move api surface symbol tests
